Strategy_02/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
requests>=2.27.0
pytz>=2022.1
schedule>=1.1.9
python-dotenv>=0.19.0

# Dhan API
dhanhq>=1.0.0
